Fixed a calculation bug in the game configuration

+++ src/Billapong.GameConsole/Configuration/GameConfiguration.cs
@@ -30,7 +30,7 @@
         /// <summary>
         /// The hole diameter
         /// </summary>
-        public const double HoleDiameter = GameWindowWidth / GameGridSize;
+        public const double HoleDiameter = GameWindowWidth / (double)GameGridSize;
 
         /// <summary>
         /// The ball diameter
